rename existing `error` module to `wasm_error`
